Relocate sources to separate file
