+++ app/meticulous/__main__.py
@@ -5,18 +5,13 @@ python -m meticulous
 """
 from __future__ import absolute_import, division, print_function
 
-import re
-
 import click
-import urllib3
+
+from meticulous._sources import obtain_sources
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 
 __version__ = "0.1"
-
-SOURCE_MARKDOWN_URLS = [
-    "https://raw.githubusercontent.com/vinta/awesome-python/master/README.md"
-]
 
 
 @click.group(context_settings=CONTEXT_SETTINGS, invoke_without_command=True)
@@ -42,47 +37,7 @@ def run_invocation():
     """
     Execute the invocation
     """
-    for url in SOURCE_MARKDOWN_URLS:
-        check_url(url)
-
-
-def check_url(url):
-    """
-    Download and process the
-    """
-    for link in get_all_markdown_github_links(url):
-        print(link)
-
-
-def get_all_markdown_github_links(url):
-    """
-    Obtain and filter markdown links to repositories.
-    """
-    links = get_all_markdown_links(url)
-    for link in links:
-        mobj = re.match(
-            "https://github.com/([A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+)(?:/|$)", link
-        )
-        if mobj:
-            yield mobj.group(1)
-
-
-def get_all_markdown_links(url):
-    """
-    Obtain all the markdown links in the URL
-    """
-    data = download_url(url)
-    matches = re.findall("[(]([^)]+)[)]", data)
-    return matches
-
-
-def download_url(url):
-    """
-    Obtain the URL content
-    """
-    http = urllib3.PoolManager()
-    resp = http.request("GET", url)
-    return resp.data.decode("utf-8")
+    print(list(obtain_sources()))
 
 
 if __name__ == "__main__":
